feat: creacion de excepciones personalizadas
se crean excepciones personalizadas

+++ src/main/java/com/inventarios/ServiceImpl/ProductoServiceImpl.java
@@ -4,53 +4,56 @@
  */
 package com.inventarios.ServiceImpl;
 
+import Exceptions.ProductoNotFoundException;
 import com.inventarios.Repository.ProductoRepository;
 import com.inventarios.Service.IProductoService;
 import com.inventarios.modelo.Producto;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.parsing.Problem;
 import org.springframework.stereotype.Service;
+
 /**
  *
  * @author ANTONIO
  */
 @Service
-public class ProductoServiceImpl implements IProductoService{
-    
+public class ProductoServiceImpl implements IProductoService {
+
     @Autowired
     private ProductoRepository productoRepositorio;
-    
 
     @Override
     public List<Producto> listarProductos() {
-       
-           return this.productoRepositorio.findAll();
+
+        return this.productoRepositorio.findAll();
     }
 
     @Override
     public Producto BuscarProductoById(Integer idProducto) {
-       
-           Producto producto = this.productoRepositorio.findById(idProducto).orElseThrow(null);
-           
-           return producto;
+
+        Producto producto = this.productoRepositorio.findById(idProducto).orElseThrow(null);
+
+        return producto;
     }
 
     @Override
     public Producto guardarProducto(Producto producto) {
-         
-          Producto newProducto= this.productoRepositorio.save(producto);
-          
-          return newProducto;
+
+        Producto newProducto = this.productoRepositorio.save(producto);
+
+        return newProducto;
     }
 
-   @Override
+    @Override
     public void deleteProducto(Integer idProducto) {
-    Producto producto = this.productoRepositorio.findById(idProducto)
-        .orElseThrow(() -> new RuntimeException("Producto con ID " + idProducto + " no existe"));
-    
-    this.productoRepositorio.deleteById(idProducto);
-     }
-    
-    
+        Optional<Producto> producto = this.productoRepositorio.findById(idProducto);
+
+        if (producto.isEmpty()) {
+            throw new ProductoNotFoundException("El producto con ID " + idProducto + " no fue encontrado.");
+        }
+
+        this.productoRepositorio.deleteById(idProducto);
+    }
+
 }
